// src/app/core/models/share/user-profile.model.ts
import { City } from './city.model';
import { District } from './district.model';
import { PaymentPlan } from './payment-plan.model';

export class UserProfile {
  accountId?: number;
  id?: number;
  email?: string;
  pin?: number;
  phoneNumber?: number;
  clubNameEn?: string;
  clubNameVi?: string;
  facebookId?: string;
  googleId?: string;
  activated?: number;
  createdAt?: string;
  createdBy?: string;
  updatedAt?: string;
  updatedBy?: string;
  deletedAt?: string;
  deletedBy?: string;
  lastName?: string;
  firstName?: string;
  identityNumber?: number;
  birthday?: string;
  avatar?: string;
  gender?: boolean;
  cityId?: number;
  clubId?: number;
  clubIds?: number[];
  clubKey?: string;
  clubKeys?: string[];
  districtId?: number;
  countryId?: number;
  roleId?: number;
  ocupationId?: string;
  accountType?: string;
  lastLogin?: string;
  firstLogin?: string;
  fingerEnroll?: boolean;
  idCardUploaded?: boolean;
  fullName?: string;
  devices?: [];
  settings?: [];
  roles?: string[];
  permissions?: string[];
  systemPermissions?: string[];
  address?: string;
  cityDetail?: City;
  districtDetail?: District;
  paymentPlanDetail?: PaymentPlan;
  memberContract?: any;
  statusContract?: string;
  idCardImage?: string;
  gmId?: string;
  ptId?: string;
  ptGroupId?: string;
}